# [ld2410] Use `Deduplicator` for sensors