--- a/esphome/components/ld24xx/__init__.py
+++ b/esphome/components/ld24xx/__init__.py
@@ -1,0 +1,1 @@
+CODEOWNERS = ["@kbx81"]