# What does this implement/fix?

Improves DRYness, reducing binary size and improving performance.

Before:
```
RAM:   [=         ]  14.8% (used 48600 bytes from 327680 bytes)
Flash: [========= ]  91.2% (used 1673254 bytes from 1835008 bytes)
```
After:
```
RAM:   [=         ]  14.8% (used 48600 bytes from 327680 bytes)
Flash: [========= ]  91.2% (used 1673236 bytes from 1835008 bytes)
```

Using `runtime_stats`:

- Before:
```
[21:30:32][I][runtime_stats:038]: Component Runtime Statistics
[21:30:32][I][runtime_stats:039]: Period stats (last 60000ms):
[21:30:32][I][runtime_stats:060]:   ld2410: count=5514, avg=0.30ms, max=350ms, total=1680ms
[21:31:32][I][runtime_stats:060]:   ld2410: count=6913, avg=0.20ms, max=17ms, total=1376ms
[21:32:32][I][runtime_stats:060]:   ld2410: count=6913, avg=0.20ms, max=16ms, total=1358ms
[21:33:32][I][runtime_stats:060]:   ld2410: count=6909, avg=0.20ms, max=18ms, total=1409ms
[21:34:32][I][runtime_stats:060]:   ld2410: count=6912, avg=0.19ms, max=17ms, total=1347ms

[21:30:32][I][runtime_stats:066]: Total stats (since boot):
[21:30:32][I][runtime_stats:078]:   ld2410: count=5514, avg=0.30ms, max=350ms, total=1680ms
[21:31:32][I][runtime_stats:078]:   ld2410: count=12427, avg=0.25ms, max=350ms, total=3056ms
[21:32:32][I][runtime_stats:078]:   ld2410: count=19340, avg=0.23ms, max=350ms, total=4414ms
[21:33:32][I][runtime_stats:078]:   ld2410: count=26249, avg=0.22ms, max=350ms, total=5823ms
[21:34:32][I][runtime_stats:078]:   ld2410: count=33161, avg=0.22ms, max=350ms, total=7170ms
```

- After:
```
[21:36:09][I][runtime_stats:038]: Component Runtime Statistics
[21:36:09][I][runtime_stats:039]: Period stats (last 60000ms):
[21:36:09][I][runtime_stats:060]:   ld2410: count=5393, avg=0.31ms, max=350ms, total=1695ms
[21:37:09][I][runtime_stats:060]:   ld2410: count=6929, avg=0.17ms, max=16ms, total=1204ms
[21:38:09][I][runtime_stats:060]:   ld2410: count=6934, avg=0.17ms, max=14ms, total=1199ms
[21:39:09][I][runtime_stats:060]:   ld2410: count=6926, avg=0.18ms, max=15ms, total=1230ms
[21:40:09][I][runtime_stats:060]:   ld2410: count=6932, avg=0.17ms, max=15ms, total=1160ms

[21:36:09][I][runtime_stats:066]: Total stats (since boot):
[21:36:09][I][runtime_stats:078]:   ld2410: count=5393, avg=0.31ms, max=350ms, total=1695ms
[21:37:09][I][runtime_stats:078]:   ld2410: count=12322, avg=0.24ms, max=350ms, total=2899ms
[21:38:09][I][runtime_stats:078]:   ld2410: count=19256, avg=0.21ms, max=350ms, total=4098ms
[21:39:09][I][runtime_stats:078]:   ld2410: count=26182, avg=0.20ms, max=350ms, total=5328ms
[21:40:09][I][runtime_stats:078]:   ld2410: count=33114, avg=0.20ms, max=350ms, total=6488ms
```


## Types of changes

- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] Code quality improvements to existing code or addition of tests
- [ ] Other

**Related issue or feature (if applicable):**

- fixes <link to issue>

**Pull request in [esphome-docs](https://github.com/esphome/esphome-docs) with documentation (if applicable):**

- esphome/esphome-docs#<esphome-docs PR number goes here>

## Test Environment

- [ ] ESP32
- [x] ESP32 IDF
- [ ] ESP8266
- [ ] RP2040
- [ ] BK72xx
- [ ] RTL87xx

## Example entry for `config.yaml`:

```yaml
# Example config.yaml

```

## Checklist:
  - [x] The code change is tested and works locally.
  - [ ] Tests have been added to verify that the new code works (under `tests/` folder).

If user exposed functionality or configuration variables are added/changed:
  - [ ] Documentation added/updated in [esphome-docs](https://github.com/esphome/esphome-docs).